requirements.txt
streamlit==1.32.2
openai==1.99.3
python-dotenv==1.0.1
Pillow==10.3.0
jiter>=0.4.0